[TimingAssistant] Improve line linking
Line linking turned out to be much more complex than I imagined at
first. Since we're changing the end line of previous line with respect
to start time of current line, we need to make sure that the end line of
previous line is not close to keyframe. So I added few checks to
accommodate that.

Plus a few small changes here and there.